scraper/requirements.txt
amqp==5.2.0
attrs==23.2.0
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
exceptiongroup==1.2.1
fake-useragent==1.5.1
Faker==25.3.0
filelock==3.13.4
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.2.0
jmespath==1.0.1
kombu==5.3.7
lxml==5.2.1
outcome==1.3.0.post0
packaging==24.0
parsel==1.9.1
prompt-toolkit==3.0.43
Protego==0.3.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
Scrapy==2.11.1
scrapy-fake-useragent==1.4.4
scrapy-rotating-proxies==0.6.2
scrapy-selenium==0.0.7
selenium==4.19.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tldextract==5.1.2
trio==0.25.0
trio-websocket==0.11.1
Twisted==24.3.0
typing==3.7.4.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
w3lib==2.1.2
wcwidth==0.2.13
webdriver-manager==4.0.1
wsproto==1.2.0
zope.interface==6.3
